backend/apps/users/serializers.py
from rest_framework import serializers

from .models import User


class UserSerializer(serializers.ModelSerializer):
    """Сериализатор пользователя."""
    class Meta:
        model = User
        fields = ('id', 'email', 'role', 'telegram_username')
        read_only_fields = fields
